+++ src/store/mod.rs
@@ -1,7 +1,7 @@
-pub mod memory;
-
 mod store_trait;
 pub use store_trait::*;
+
+pub mod memory;
 
 #[cfg(feature = "postgres-store")]
 pub mod postgres;
@@ -11,3 +11,8 @@ pub mod redis;
 
 #[cfg(feature = "layered-store")]
 pub mod layered;
+
+#[cfg(feature = "layered-store")]
+mod layered_store_trait;
+#[cfg(feature = "layered-store")]
+pub use layered_store_trait::*;